Restructure app using service configs
