--- src/database/mod.rs
@@ -1,9 +1,0 @@
-//! src/database/mod.rs
-
-pub mod database;
-pub mod mock;
-pub mod tables;
-
-pub use database::*;
-pub use mock::*;
-pub use tables::*;
