--- src/graphql/mod.rs
@@ -1,9 +1,0 @@
-//! src/graphql/mod.rs
-
-pub mod mutations;
-pub mod queries;
-pub mod schema;
-
-pub use mutations::*;
-pub use queries::*;
-pub use schema::*;
