+++ src/services/graphql/mod.rs
@@ -1,0 +1,9 @@
+//! src/services/graphql/mod.rs
+
+mod config;
+mod routes;
+mod schema;
+
+pub use config::*;
+pub use routes::*;
+pub use schema::*;
