+++ src/services/graphql/schema/context.rs
@@ -1,0 +1,5 @@
+//! src/services/graphql/context.rs
+
+use crate::utils::database::DatabaseConnection;
+
+impl juniper::Context for DatabaseConnection {}
