+++ src/services/graphql/schema/subscriptions.rs
@@ -1,0 +1,5 @@
+//! src/services/graphql/subscriptions.rs
+
+use crate::utils::database::DatabaseConnection;
+
+pub type Subscriptions = juniper::EmptySubscription<DatabaseConnection>;
