+++ src/services/mod.rs
@@ -1,0 +1,7 @@
+//! src/services/mod.rs
+
+pub mod graphql;
+pub mod rest;
+
+pub use graphql::*;
+pub use rest::*;
