+++ src/services/rest/config.rs
@@ -1,0 +1,3 @@
+//! src/services/rest/config.rs
+
+// TODO.
