+++ src/services/rest/mod.rs
@@ -1,0 +1,5 @@
+//! src/services/rest/mod.rs
+
+mod config;
+
+pub use config::*;
