+++ src/utils/database/mod.rs
@@ -1,0 +1,7 @@
+//! src/utils/database/mod.rs
+
+pub mod connection;
+pub mod schema;
+
+pub use connection::*;
+pub use schema::*;
