+++ src/utils/mocks/mod.rs
@@ -1,0 +1,7 @@
+//! src/utils/mocks/mod.rs
+
+pub mod requisition;
+pub mod requisition_line;
+
+pub use requisition::*;
+pub use requisition_line::*;
