+++ src/utils/mod.rs
@@ -1,0 +1,7 @@
+//! src/utils/mod.rs
+
+pub mod database;
+pub mod mocks;
+
+pub use database::*;
+pub use mocks::*;
